Design review - button adjustments
- Renamed WKResizableButton to WKSmallButton
- Added quiet and neutral configurations to match design components guidelines
- Used neutral configuration in WKEmptyView
- Deleted WKSecondaryButton
- Used quiet configuration in WKOnboardingView learn more button
