+++ Components/Sources/Components/Components/Empty View/WKEmptyView.swift
@@ -39,7 +39,7 @@ public struct WKEmptyView: View {
                                 .multilineTextAlignment(.center)
                         }
                         if type == .noItems {
-                            WKResizableButton(configuration: .medium, title: viewModel.localizedStrings.buttonTitle, action: delegate?.emptyViewDidTapSearch)
+                            WKSmallButton(configuration: .neutral, title: viewModel.localizedStrings.buttonTitle, action: delegate?.emptyViewDidTapSearch)
                                 .padding([.leading, .trailing], 32)
                         }
                         Spacer()
